+++ CHANGELOG.md
@@ -10,3 +10,10 @@ Version 1.0.1 *(2017-12-18)*
 ----------------------------
 
 - Fix Jitpack build issues with 1.0.0.
+
+Version 1.0.2 *(2018-01-05)*
+----------------------------
+
+- Add Travis CI
+- Fix wrong data and view when scrolled - merge pull request
+- Modify ` fun addItems(vararg items: T)` to include index as `fun addItems(vararg items: T, index: Int = itemCount)`